% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterProfilerTable.R
\name{ClusterProfilerTable-class}
\alias{ClusterProfilerTable-class}
\alias{ClusterProfilerTable}
\alias{.fullName,ClusterProfilerTable-method}
\alias{.panelColor,ClusterProfilerTable-method}
\alias{.generateTable,ClusterProfilerTable-method}
\alias{.hideInterface,ClusterProfilerTable-method}
\title{clusterProfiler Result Table}
\description{
A table that dynamically computes pathway analysis results for a given gene set.
}
\section{Slot overview}{

This class does not define any additional slots yet.

This class inherits all slots from its parent
\code{Table-class} and \code{Panel-class} classes.
}

\section{Constructor}{

\code{ClusterProfilerTable(...)} creates an instance of a ClusterProfilerTable class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{ClusterProfilerTable} class.
Refer to the documentation for each method for more details on the remaining arguments.

For defining the interface:
\itemize{
\item \code{.fullName(x)} will return \code{"clusterProfiler Table"}.
\item \code{.hideInterface(x)} will return \code{TRUE} for UI elements related to multiple column selections,
otherwise calling the method for \linkS4class{RowTable}.
\item \code{.panelColor(x)} will return the specified default color for this panel class.
}

For creating the table:
\itemize{
\item \code{.generateTable(x, envir)} will create a data.frame of newly computed statistics in \code{envir}.
The method will return the commands required to do so.
}
}

