% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{iSEEOptions}
\alias{iSEEOptions}
\title{Default panel options}
\format{
An object of class \code{list} of length 4.
}
\usage{
iSEEOptions
}
\description{
Set global options using \code{iSEEOptions$set()}, so that all relevant panels will use these default values during initialization.
This function mergees
}
\details{
See \code{str(iSEEOptions$get())} for a list of default panel options.

Note that \code{iSEEOptions$restore()} can be used to reset the global options to the package default values.
}
\section{Available options}{


\describe{
\item{\code{orgdb.package}}{Description.}
\item{\code{orgdb.key}}{Description.}
\item{\code{orgdb.keytype}}{Description.}
}
}

\examples{
iSEEOptions$get('downsample'); iSEEOptions$get('org.db')
}
\author{
Kevin Rue-Albrecht
}
\keyword{datasets}
